Fix Weed Nodes in Walls (#3765)
# About the pull request

This PR makes it so weed nodes cannot be placed in a turf with density
set true (such as when you build a resin wall on top of yourself) or a
dense structure (such as a window frame). Nodes still can be placed
under doors while they are open. Marked it as balance but feel free to
change it to fix if you think it isn't balance.

If theres a situation where a weed node should be able to be placed in a
turf with density let me know and I can make the check exclusively for
resin walls. I can't think of any though.

# Explain why it's good for the game

Fixes #3745
Fixes #3770 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/89b16684-1f35-40f6-81f0-44855f28e732)

</details>

# Changelog
:cl: Drathek
balance: Weed nodes can no longer be placed in walls or window frames
(or any turf or structure with density)
refactor: Refactored the plant weeds ability code
/:cl: