Working Joes no longer can have a gradient on their rare hair. (#3827)
# About the pull request

Resolves oversight that allows for a user to have their own hair color
as a Working Joe by enabling Gradient in character select, to avoid
accidently overriding the default hair color. Fixes #3820

# Explain why it's good for the game

While WJs are whitelisted, this mechanical enforcement is more so to
prevent accidents from when people join as Working Joe without removing
gradients, as a person should not need to design a character to fit a
role that should overwrite that in the first place.

# Changelog
:cl:
fix: Working Joes can no longer have a gradient on their rare hair
spawn.
/:cl: